// homebridge/interfaces/device.ts
export interface Device {
    dvcAuthCd: string
    rq_stat_cd: string
    dvcRegStatCd: string
    dvcModel: string
    ordNo: string
    dvcTypeCd: string
    mbrDvcSeq: number
    regDttm: string
    releaseDttm: string
    prodType: string
    selfManageYn: string
    dvcSeq: number
    comType: string
    dvcBrandCd: string
    dvcAuthReasonCd: string
    guideTypeCd: string
    instYn: string
    sellTypeCd: string
    stationCd: string
    admdongCd: string
    membershipYn: string
    registUserCount: number
    installAddr: string
    netStatus: boolean
    pairingDttm: string
    iotYn: string
    isGubun: string
    prodNameFull: string
    buyerMbrSeq: string
    roomPos: string
    rqDttm: string
    prodCd: string
    salesCntryId: string
    userGradeCd: string
    unreadYN: string
    prodName: string
    heartsvcDttm: string
    userTypeCd: string
    wifiType: string
    sort: number
    dvcNick: string
    resetDttm: string
    instDttm: string
    barcode: string
    gpsLat: string
    gpsLng: string
    gpsNationCd: string
    eUseTp: string
}